Simplifying refactor
I figured the formula punctuation are more of an aid to the
developer than a part of Nock in the wild, which is basically just a
number soup, so I got rid of the reification into the Formula type.

Also ran into a weird Rust pattern matching bug I need to work
around with the trivial tuple in the fn nock.